Resolving various requested changes from PR

+++ DataLayer/CmuData/CmuData.cpp
@@ -11,5 +11,5 @@ CmuData::~CmuData()
 
 I_CmuUnit& CmuData::getCmuUnit(const unsigned char& index)
 {
-    return cmuUnits_[static_cast<int>(index)];
+    return cmuUnits_[index];
 }
